move code out ofinterpreter file and make it download and load code during initialization
